+++ Example/CustomizedTokenViewController.swift
@@ -37,6 +37,8 @@ class CustomizedTokenViewController: UIViewController, ICTokenFieldDelegate {
   override func loadView() {
     super.loadView()
     view.backgroundColor = UIColor.white
+    textView.isEditable = false
+    textView.isSelectable = false
     textView.text = "[\n\n]"
     textView.font = UIFont.preferredFont(forTextStyle: .subheadline)
     textView.frame = view.bounds.insetBy(dx: 10, dy: 10)
@@ -83,14 +85,27 @@ class CustomizedTokenViewController: UIViewController, ICTokenFieldDelegate {
     print(#function)
   }
 
-  func tokenField(_ tokenField: ICTokenField, didEnterText text: String) {
-    print("Add: \"\(text)\"")
+  func tokenField(_ tokenField: ICTokenField, didChangeInputText text: String) {
+    print("Typing \"\(text)\"")
+  }
+
+  func tokenField(_ tokenField: ICTokenField, shouldCompleteText text: String) -> Bool {
+    print("Should add \"\(text)\"?")
+    return text != "42"
+  }
+
+  func tokenField(_ tokenField: ICTokenField, didCompleteText text: String) {
+    print("Added \"\(text)\"")
     updateTexts()
   }
 
   func tokenField(_ tokenField: ICTokenField, didDeleteText text: String, atIndex index: Int) {
-    print("Delete: \"\(text)\"")
+    print("Deleted \"\(text)\"")
     updateTexts()
+  }
+
+  func tokenField(_ tokenField: ICTokenField, subsequentDelimiterForCompletedText text: String) -> String {
+    return " ,"
   }
 
   // MARK: - UIResponder Callbacks