+++ Source/TokenField/ICTokenField.swift
@@ -34,10 +34,16 @@ import UIKit
   @objc optional func tokenFieldDidEndEditing(_ tokenField: ICTokenField)
   /// Tells the delegate that the token field will process the pressing of the return button.
   @objc optional func tokenFieldWillReturn(_ tokenField: ICTokenField)
+  /// Tells the delegate the input text is changed.
+  @objc optional func tokenField(_ tokenField: ICTokenField, didChangeInputText text: String)
+  /// Asks the delegate if the text should become a token in the token field.
+  @objc optional func tokenField(_ tokenField: ICTokenField, shouldCompleteText text: String) -> Bool
   /// Tells the delegate that the text becomes a token in the token field.
-  @objc optional func tokenField(_ tokenField: ICTokenField, didEnterText text: String)
+  @objc optional func tokenField(_ tokenField: ICTokenField, didCompleteText text: String)
   /// Tells the delegate that the token at certain index is removed from the token field.
   @objc optional func tokenField(_ tokenField: ICTokenField, didDeleteText text: String, atIndex index: Int)
+  /// Asks the delegate for the subsequent delimiter string for a completed text in the token field.
+  @objc optional func tokenField(_ tokenField: ICTokenField, subsequentDelimiterForCompletedText text: String) -> String
 }
 
 
@@ -287,23 +293,28 @@ open class ICTokenField: UIView, UITextFieldDelegate, ICBackspaceTextFieldDelega
     }
 
     let text = (input as NSString).replacingCharacters(in: range, with: string)
+    delegate?.tokenField?(self, didChangeInputText: text)
 
     for delimiter in delimiters {
-      if text.hasSuffix(delimiter) {
-        let index = text.index(text.endIndex, offsetBy: -delimiter.characters.count)
-        let newToken = text.substring(to: index)
-        textField.text = nil
-
-        if !newToken.isEmpty && newToken != delimiter {
-          tokens.append(ICToken(text: newToken, normalAttributes: normalTokenAttributes, highlightedAttributes: highlightedTokenAttributes))
-          layoutTokenTextField()
-          delegate?.tokenField?(self, didEnterText: newToken)
-        }
-        togglePlaceholderIfNeeded()
-
-        return false
+      guard text.hasSuffix(delimiter) else {
+        continue
       }
+
+      let index = text.index(text.endIndex, offsetBy: -delimiter.characters.count)
+      let newText = text.substring(to: index)
+
+      if !newText.isEmpty && newText != delimiter && (delegate?.tokenField?(self, shouldCompleteText: newText) ?? true) {
+        tokens.append(customizedToken(with: newText))
+        layoutTokenTextField()
+        delegate?.tokenField?(self, didCompleteText: newText)
+      }
+
+      textField.text = nil
+      togglePlaceholderIfNeeded()
+
+      return false
     }
+
     return true
   }
 
@@ -316,7 +327,7 @@ open class ICTokenField: UIView, UITextFieldDelegate, ICBackspaceTextFieldDelega
 
   // MARK: - ICBackspaceTextFieldDelegate
 
-  func textFieldShouldDelete(_ textField: ICBackspaceTextField) -> Bool {
+  @nonobjc func textFieldShouldDelete(_ textField: ICBackspaceTextField) -> Bool {
     if tokens.isEmpty {
       return true
     }
@@ -363,6 +374,14 @@ open class ICTokenField: UIView, UITextFieldDelegate, ICBackspaceTextFieldDelega
   }
 
   // MARK: - Private Methods
+
+  private func customizedToken(with text: String) -> ICToken {
+    if let string = delegate?.tokenField?(self, subsequentDelimiterForCompletedText: text) {
+      return ICToken(text: text, delimiter: string, normalAttributes: normalTokenAttributes, highlightedAttributes: highlightedTokenAttributes)
+    } else {
+      return ICToken(text: text, normalAttributes: normalTokenAttributes, highlightedAttributes: highlightedTokenAttributes)
+    }
+  }
 
   /// Returns true if any highlighted token is found and removed, otherwise false.
   private func removeHighlightedToken() -> Bool {
@@ -436,10 +455,16 @@ open class ICTokenField: UIView, UITextFieldDelegate, ICBackspaceTextFieldDelega
     guard let text = inputTextField.text, !text.isEmpty else {
       return
     }
+
+    let shouldCompleteText = delegate?.tokenField?(self, shouldCompleteText: text) ?? true
+    guard shouldCompleteText else {
+      return
+    }
+
     inputTextField.text = nil
-    tokens.append(ICToken(text: text, normalAttributes: normalTokenAttributes, highlightedAttributes: highlightedTokenAttributes))
+    tokens.append(customizedToken(with: text))
     layoutTokenTextField()
-    delegate?.tokenField?(self, didEnterText: text)
+    delegate?.tokenField?(self, didCompleteText: text)
   }
 
   /// Removes the input text and all displayed tokens.